Refactor: Refactor rest of components
Footer, SVGIconBtn, UserAvatar
